python-grpc-server/requirements.txt
matplotlib~=3.7.1
grpcio
grpcio-tools

numpy~=1.24.3